[Feature] Remove Customer Support & Fixed Bug in Show All Customers

+++ Customer.h
@@ -5,7 +5,7 @@ typedef struct {
     char* id;            // Unique identifier for the customer
     char* name;              // Customer's name
     char* joinDate;           // Date the customer joined (format: DD-MM-YYYY)
-    float totalAmountSpent;  // Total amount the customer has spent
+    double totalAmountSpent;  // Total amount the customer has spent
     int itemsPurchased;       // IDs of up to 3 items the customer purchased
 } Customer;
